requirements.txt
altgraph==0.17.2
bcrypt==3.2.2
cffi==1.15.1
coverage==6.4.2
cryptography==37.0.4
cycler==0.11.0
distlib==0.3.5
et-xmlfile==1.1.0
Everything-Tkinter==0.2
filelock==3.7.1
fonttools==4.34.4
future==0.18.2
ifaddr==0.2.0
iso8601==1.0.2
kiwisolver==1.4.4
matplotlib==3.5.2
numpy==1.23.1
openpyxl==3.0.10
packaging==21.3
paramiko==2.11.0
pefile==2022.5.30
Pillow==9.2.0
platformdirs==2.5.2
pycparser==2.21
pyinstaller==5.3
pyinstaller-hooks-contrib==2022.8
PyNaCl==1.5.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyserial==3.5
python-dateutil==2.8.2
pywin32-ctypes==0.2.0
PyYAML==6.0
robotframework==5.0.1
six==1.16.0
virtualenv==20.16.2
pymodbus==3.0.0
